# refactor: parallel processing