### What:

- [ ] Bug Fix
- [ ] New Feature

### Description:

Implemented parallel processing in the Aspell Spellchecker class, resulting in a 40% performance improvement on the first (non-cached) run.  

### Testing

To more easily test this, manually comment out caching in the Aspell class, run Pest while switching between main branch and this branch.

#### Test notes

**Main branch** with caching turned off
- 12.22s
- 11.97s
- 11.98s

**This branch** with caching turned off
- 8.20s
- 8.26s
- 8.24s