@chr15k 2025-01-18
> It may be worth adding a test if possible?

Hey @benjam-es, after adding more tests, I noticed that performance worsens with large datasets (e.g., if $text was 10,000+ lines), as it spawns n processes. I then tried chunking, which helped but was still slower than the main branch for this size. Given this, I actually don’t think it’s worth committing since the change isn't significant enough an improvement and adds extra complexity. I think if we can improve the first-run speed we'll need a better/different approach. Thanks for looking!